Update patch set 1
Patch Set 1:

I see you haven't +1'd yet, but I thought I could provide some early feedback. I hadn't really understood the implications before or else I would have raised them earlier.

This seems fine enough if there is an actual use case. But I don't know if it ever will be used and what I worry is that we are making API at the JNI level that we will then be stuck with because we don't know if someone may be using it. To mitigate against this at a minimum I would add a comment on the function asking that if someone provides their own native layer that they feedback that fact. The "API" should be marked as provisional. Finally, it is important to note that in effect by doing this we are saying that the JNI layer is legitimate way to extend this plug-in, so it is not just the new method that becomes API, but the entire JNI interface.

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>