Update patch set 1
Patch Set 1:

> If I list the content of https://download.eclipse.org/staging/2020-12/plugins/ with a browser, I can see version 6.1.0 of org.eclipse.cdt.core.native, so I don't understand why maven can't find it.
> Any idea?

It is almost certainly the cache on download.eclipse.org - there is a fundamental problem with how it is designed, it assumes that files are added and not edited, but of course in the case of artifacts/contents files they are often edited. This is normally resolved by waiting until the cache expires, but sometimes the webmaster seems to have to manually dump the cache.

The EPP project, which is the "official" consumer of the staging repo actually accesses the repo via file rather than http access, so the EPP build is never affected.

The content/artifact files are saved in ~/.m2/repository/.cache/tycho/p2-repository-metadata/  - but I don't know how the mapping works.

I have raised it as part of https://bugs.eclipse.org/bugs/show_bug.cgi?id=568117#c2

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>