Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> @Jonah Do you plan to update CDT.setup as well?

I don't think there is anything to update in cdt.setup as it does not explicitly say which bundles to include from repos.

One significant challenge with this set of changes is that p2 does not resolve "uses" and as a result a solution that p2 comes up with may not work for OSGi.

Do you have time to test this before I merge it (once it builds)? Right now this is a blocking change because our orbit repo changed under our feet (I kind of want to go back to explicit orbit repos as this is the second time in 2020-12 that our builds have been in bad shape because we get orbit updates implicitly, but platform updates explicitly).

Patch-set: 1