Update patch set 3
Patch Set 3:

(2 comments)

> Patch Set 3: Code-Review+1
> 
> Finally I can mute it!

We might consider dropping the patch for bug 486033 until a cleaner solution is found (one that changes also the life-cycle of the consoles, so that adding the close button is not bug-prone).

Even with the close button the preference will need to be use to prevent the console from popping up again. So the close button will be good in cases were you do a long build and you close the console as you won't do another CDT build for a long time (which is probably not the most common scenario).

Patch-set: 3