Update patch set 3
Patch Set 3:

> What exactly is bug-prone?

BuildConsoleManager caches values for the entire Eclipse session. In my suggested fix for bug 486033 that doesn't change; we only recreate the build console and re-add it in case its requested after being closed. I doubt that is robust; if it doesn't introduce bugs now, it might introduce bugs in future.

Patch-set: 3