Update patch set 6
Patch Set 6:

I think patch set 6 makes the change more or less minimal and the hack-like aspect is hopefully negligible.

Whatever internal state is cached in BuildConsolePage remains, the only change is re-adding the console to the console view whenever its requested (and was closed prior to the request).

Patch-set: 6