Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (1 comment)
> 
> > Patch Set 4: Code-Review+1
> > 
> > Should be ready to merge.
> 
> There are unresolved comments and an open question from Andrey:
> 
> > One step further, but still unclear to me what do you mean. *Which* "values" are problematic and how that relates to the closed console?

Sorry I'm not seeing the unresolved comments. Which ones did I miss?

The question from Andrey is related to the change for bug 486033; it relates to this change in that this change might be sufficient to not bother with bug 486033 at all.

Patch-set: 4
Reviewer: Gerrit User 119988 <119988@97ee7c02-f12f-4043-b43e-dea463d88b31>