Update patch set 2
Patch Set 2:

The conclusion that the CustomizePerspectiveDialog is broken is correct. If everything was well, we would be just queering the E4 model instead of this complex population. We already do that for the toolbar items, see Bug 549898.

However, the E4 model and the items that are showed via the MenuManagers are not in sync. So just changing the CPD code to query the model will most likely result in an incorrect contents of the CPD. I have been working to get the menu in sync, however that is a complex task with many side effects and corner cases. I expect that a major redesign in how the E3 and E4 menu items are handled is needed. I don't expect a fix here soon, just don't have enough spare time to look into that.

So either chose for option 1 or 3. In case of option 1, you are not really going to E4 menus (that would mean u use a model fragment). You would be using the org.eclipse.ui.menus extension point introduced in 3.3. Although, if I remember correctly, some functionality might still be missing see e.g. Bug 414612.

Patch-set: 2
Reviewer: Gerrit User 4326 <4326@97ee7c02-f12f-4043-b43e-dea463d88b31>