Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> OK, talked to Andrey. I'll try to replace the 3 actions with commands/menus (one of the actions has a submenu I believe). Might take a bit of time though.

Thanks to all of you for investing time so far into the investigation.

I assume that this is no longer in the 2020-12 time frame? Can you move this change to work in progress state please? Or abandon it if it is not relevant.

Patch-set: 2
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>