Update patch set 2
Patch Set 2:

I think with patch set 2 the commands have the correct enablement expressions (same as the previous actions).

So please review; I don't think we want this change in the soon-to-be-released CDT though. Better that its added at the start of development of a new version. So that if there are problems, those are found before an actual release.

Patch-set: 2