Update patch set 1
Patch Set 1:

(7 comments)

> Patch Set 1:
> 
> (7 comments)
> 
> Thank you for this excellent piece of work. I like how much improved this will make the code. There are a few minor comments/typos to have a look at if you can.
> 
> I am quite curious about your use case that led you to improve this code? You can send me a private email if you can't discuss it here, but I am always interested to learn who and how CDT is being extended.

Yes, I've tried to address the comments. We were trying to use the existing SyncUtil for our custom integration test cases as the API it provides is pretty close to what we require. We extended it and made our own SyncUtil to add some custom APIs but for the test engineers to get comfortable with the existing ones we wanted to have some javadoc

I believe the existing test framework in CDT is quite nice and users can extend it a bit to suite their use cases

Patch-set: 1