Update patch set 2
Patch Set 2:

(4 comments)

> Yes, I've tried to address the comments. We were trying to use the existing SyncUtil for our custom integration test cases as the API it provides is pretty close to what we require. We extended it and made our own SyncUtil to add some custom APIs but for the test engineers to get comfortable with the existing ones we wanted to have some javadoc

Excellent, thanks for the background.

> I believe the existing test framework in CDT is quite nice and users can extend it a bit to suite their use cases

Great. In case you didn't know, please do keep in mind that while SyncUtil hasn't changed much in the last years, and is unlikely to going forward, the code is not actually API and is therefore subject to change without notice or requiring major or minor version changes.

I have marked as resolved all the comment threads I could. There are still a couple of open ones. Please let me know when you are done by adding a Code-Review +1 to the change.

Patch-set: 2