Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> This contribution is over 1000 lines of code and a CQ may be needed[1].
> Project committers, please confirm if a CQ is needed by replying to this comment (in Gerrit) leaving the quote intact, and adding Verified +1, and Genie will automagically create a CQ.
> 
> [1] https://eclipse.org/legal/EclipseLegalProcessPoster.pdf.
> -genie:cq

Thanks Michael for all the tests. As this is a substantial change there are a few checks that need to be run. As long as you sign off the contribution (git commit --amend --sign-off) and add a copyright header to the new file then there isn't much else for you to do. If any of the code was copied from elsewhere, please add that to the commit message.

Patch-set: 3