Bug 569581 - Rethrow IOException in ElfParser.hasInterpProgramHeader()
This change adjust behavior in ElfParser.hasInterpProgramHeader() (added
for bug 512822) to throw an IOException that occurred, as opposed to
logging the exception and continuing.

An IOException indicates an I/O problem with the binary. E.g. the binary
can be deleted by the user or by Eclipse, while being visited in
CModelManager. If hasInterpProgramHeader() returns instead of throwing,
more problems can occur "later on".

Change-Id: I8852d9cbaa17dc97f668bb1666e9b046bbde1fca
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>