Bug 516371: Enable GDB target async support for Windows remote debugging
Traditionally CDT used sync debug support all the time. However there
are cases where using target async is better because of GDB missing
interrupt.

This patch expands the cases that use target async to be all
remote targets when using Windows host. That is in addition to
cases when the full GDB Console is supported (new-ui).

Signed-off-by: Muhammad Bilal <hafizbilal100@gmail.com>
Change-Id: I1309d20319a24e4f23543d4ed22735044fd5b811