WIP: Identify what may be broken in core tests if run without suite
This change renames all suites to be called ...TestSuite so that
the default includes don't pick them up. With the name ...Tests it
means that those tests referenced by the suite are run multiple times.

Tests which relied on being run as part of a TestSuite have been
refactored to run as individual tests by moving the logic from
the suite into the test (e.g. CommentHandlingTestSuite)

Tests that were not runnable have been renamed from ...Test to ...Tester
to match a convention already used in CDT.

Change-Id: Ia91e12fe76c3ec2d914463a28400d21b9daf1910