Update patch set 5
Patch Set 5:

PS I am on holidays, but I have some time to look at this. Thanks for your patience on the review. On a quick look this is in the right direction. Only question is what happens for targets that don't support async, like I believe windows doesn't?

Patch-set: 5