Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 5:
> > 
> > > @John @Torbjörn - Am I missing something here - why is this problem only surfacing now? Has interrupting remote targets from Windows been a long standing issue that I just wasn't aware of???
> > 
> > To be clear, I think that using async is a better solution in this use case. It definitely fixes a problem
> 
> @Jonah, I have not experienced the problem which this patch is intended to solve, but I have no reason to doubt the diagnosis. The patch should bring support for remote target debugging on Windows hosts more in line with remote target debugging on Linux hosts. This seems to be a good point in the release cycle to apply the patch, allowing time to back out the change of any problems arise.

@Jonah, I agree with John on this one. I have used both all-stop and non-stop modes in GDB for quite a while in the products that I've been working on an never faced this issue, at least not that I can recall.
If the solution is to apply this gerrit, then what should we do about earlier versions of GDB that obviously will still have this weakness?
Is it time to ditch support for earlier version of GDB than 8.0 in CDT? GDB 8.0 was released 2017-06-04, so almost 3,5 years ago.

Regarding the change it self; I think the code changes looks clean but I would like to get the commit message a little mote condensed or at least easier to read. It's a bit too much text in one block for my taste and I've tried to read the full comment a few times now but looses focus and gets lost after about half of it. I also have a feeling that most of the commit message should be documented in the ticket rather than having all duplicated (bugzilla + git).

Patch-set: 6
CC: Gerrit User 178588 <178588@97ee7c02-f12f-4043-b43e-dea463d88b31>