Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I checked all call sites for this method and conclude your fix looks right. However there is a place that a getPHdrs(path) throws IOException that makes me wonder if it needs to be change to make this change ok. In the removed try/catch org.eclipse.cdt.utils.elf.Elf.getPHdrs() is eventually thrown and it throws an IOException if a header with e_ident==ELFCLASSNONE (or unknown) is read.
> 
> Simeon, have you examined that case yet?

If hasInterpProgramHeader() throws an exception, getBinary() will either ignore the exception and return a null, or it will try to call createBinaryArchive(). If that fails, the code will log the exception and re-throw (the re-thrown exception is "ignored" later on).

Without this change, hasInterpProgramHeader() will log an exception and "fail silently", then getBinary() will call createBinaryShared() and return the result. An IOException indicates something is wrong with the binary file altogether, I don't know if that behavior is correct. But with this change the behavior is definitely different.

Patch-set: 1