Update patch set 13
Patch Set 13:

Tried it out using a gtest sample.

Works fairly good, but I notice that after running tests, the Unit test view does not grab focus automatically which would be a good thing.  Similarly, if I rerun the tests, I get the Console view popping up and I have to manually click on the Unit test view.  Would be nice to always open the view after the run and give it focus.

Also, if I choose to rerun failing tests, works great.  If I then say rerun all tests, it runs the failed test again but not the others.  This seems wrong, but feel free to correct me.  If it is a UnitTest bug, then I am ok with opening a bug and still checking this in.

Patch-set: 13
Reviewer: Gerrit User 582 <582@97ee7c02-f12f-4043-b43e-dea463d88b31>