Update patch set 13
Patch Set 13:

> the Unit test view does not grab focus automatically which would be a good thing.

That is some enhancement request to report to the unit test view.

> Similarly, if I rerun the tests, I get the Console view popping up and I have to manually click on the Unit test view.  Would be nice to always open the view after the run and give it focus.

Could it be that the console has "Show console when [...] output changes" turned on and takes focus explicitly?

> Also, if I choose to rerun failing tests, works great.  If I then say rerun all tests, it runs the failed test again but not the others.  This seems wrong, but feel free to correct me.  If it is a UnitTest bug, then I am ok with opening a bug and still checking this in.

Re-run does rerun the exact previous test session. If test session was only failed tests, then only failed tests are rerun.
I don't think it's a big bug, but maybe we can improve the action description in Unit Test view to be more explicit.

Overall, it seems to me no issue is specific to that current change, and that it's fine to be merged.

Patch-set: 13