Update patch set 17
Patch Set 17:

> Patch Set 17:
> 
> > Patch Set 17:
> > 
> > > Patch Set 17:
> > > ...
> > > Am I wrong?
> > 
> > No, you are not wrong.  There should either be a single year when the file is created (e.g. 2020) or if the file is modified or copied from elsewhere, then you have the two years with the 2nd year being the last change.  If the year hasn't changed, then you do not specify double years.  In the case where you have modified the file in the same year, you can simply add a note in the changes in the header about what you did (e.g. Red Hat Inc.  - modified to use in C/C++ unit test).
> 
> @Jeff, ah. I'm sorry."2020, 2020" was in my commit, not in your change. Sorry for not spotting it earlier.

The 2020, 2020 is created by the scripts that manage this stuff in CDT and then included in your commit. The scripts in CDT are not well tested around new features being added as this is one of the first new features since the scripts were introduced.

Patch-set: 17
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>