Update patch set 1
Patch Set 1:

Thank you for your fix to CDT. There are still some flow issues to resolve on the commit. However I am happy to just handle them on this occasion. If you would like to try to do the below yourself let me know. If not I will do the below steps and submit.

Before I can submit this chain of commits, they need to be squashed into a single commit. Unlike GitHub PRs, updates in gerrit should amend the previous version so that the review stays in one gerrit, all with the same Change-Id. Also, the copyright update still needs to be done - you added a contributor entry, but didn't update the copyright year.

Patch-set: 1
CC: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>