Update patch set 6
Patch Set 6:

I am not a parsing expert but have looked at the patch and it seems reasonable, however, I can't
reproduce the problem.  When I try the example in Bug 568957 and make it a C++ Managed project and then set options for indexer/problems indexer/activity and indexer/statistics in a child Eclipse I see no error message as mentioned in the bug.  Is there something I am doing wrong to reproduce?  I have tried rebuilding and reindexing.  I added a main function and took it out to see if that was an issue.

Patch-set: 6
CC: Gerrit User 582 <582@97ee7c02-f12f-4043-b43e-dea463d88b31>