Update patch set 3
Patch Set 3:

In ClassTypeHelper.getVisibility() the objects are compared via object equality.  But in CPPVisitor.associateTypeWithValue the objects representing the types are cloned. This leads to a failure in CPPClassTypeHelper and the given stacktrace.

You introduced the code which clones the types last year (https://bugs.eclipse.org/bugs/show_bug.cgi?id=545756).

Do you know why it is necessary to associate the type with a value when the variable is const and has a numerical value?

I see multiple ways to fix my bug:

    Implement the "equals"-Method in CPPTypedef correctly, so that the cloned objects are equal to the original objects. It is a possibility to save a reference to the original object in the cloned object which would lead to good performance if the original objects are compared via Object.equals (i.e. reference comparison).
    Change the implementation of ClassTypeHelper.getVisibility, which is a bit messy as the changes have to be two places to handle using and typedef.
    Not applying the value to the type in case of const and numerical value any more, i. e. delete the lines in CPPVisitor.createType.

In number 3 these lines would be deleted:

            if (initClause instanceof IASTEqualsInitializer) {
                initClause = ((IASTEqualsInitializer) initClause).getInitializerClause();
                if (initClause instanceof IASTLiteralExpression && SemanticUtil.isConst(type)) {
                    // The check here avoids performing the computation in getValueOfInitialize()
                    // in cases where we wouldn't use it anyways because we don't have a CPPBasicType.
                    if (SemanticUtil.getNestedType(type, TDEF | ALLCVQ) instanceof CPPBasicType) {
                        type = associateTypeWithValue(type,
                                SemanticUtil.getValueOfInitializer(declarator.getInitializer(), type));
                    }
                }
            }


Although solution 3 seems tempting to me, I am not sure about the consequences. Do the enriched types help the indexer? The code should work without associating the value to the type as the type is not cloned when there is no numerical type (e.g. const char*). The next best I like solution 1.  But no other class in the package of CPPTypedef implements the "equals"-Method. Solution 2 seems messy to me.

Patch-set: 3