Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (2 comments)
> 
> > Patch Set 1:
> > 
> > Hi Torbjörn, does the CDT internal builder exhibit the same issue?
> 
> Hi John,
> 
> Thanks for your review.
> The state of this gerrit is not ready for merge yet. I just pushed what I had to have a backup and get a sense of how much work it would be to get the tests going with it.

Regarding the internal builder; I think so yes.
If I recall correctly, the limitation is either in CMD or the WinAPI, or perhaps both? It's a long time since I last had a look at this piece of code and the reasons behind it, but doing a quick google on the subject led me here: https://docs.microsoft.com/en-us/troubleshoot/windows-client/shell-experience/command-line-string-limitation

Patch-set: 1