Bug 571405: Generate rules with -MT"$@"
When there is a linked file in the project, the build target for the
content in the .d file should be the .o file as that is what is
depending on the source file.

Use "$@" rather than "$(@)" as it's the usual syntax for single letter
variables in make.

Contributed by STMicroelectronics

Change-Id: I6b024b2b3a1a8b061740e99de5e96923981c92fa
Signed-off-by: Torbjörn SVENSSON <torbjorn.svensson@st.com>