Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > @Jonah, perhaps you have an idea?
> 
> Good point - you are correct, you can't apply @file to every tool. For example https://ci.eclipse.org/cdt/job/cdt-verify-test-cdt-other-pipeline/3151/testReport/junit/org.eclipse.cdt.managedbuilder.core.tests/ManagedProject30MakefileTests/testMacroSupportInBuildDefinitions/ is a test that does a "tar" as the tool that does not understand @ files.

Would it make sense to have a boolean option on the tool that allows it to indicate if the tool supports the @file format? I guess this would be backward compatible as the default value can be set to "false" in order to only trigger this behavior when requested.
There is one problem with this solution though; what if the user changes the command in the toolchain integration configuration? Lets say that they replace "gcc" with "tar"? On the other hand, there would likely be a lot of other issues with the rest of the flags too...
Another hackish approach is to check the generated command if it ends with gcc(.exe)? and in that case, enable it.

Patch-set: 2