Update patch set 2
Patch Set 2:

> Would it make sense to have a boolean option on the tool that allows it to indicate if the tool supports the @file format? I guess this would be backward compatible as the default value can be set to "false" in order to only trigger this behavior when requested.

Sounds ok.

> There is one problem with this solution though; what if the user changes the command in the toolchain integration configuration? Lets say that they replace "gcc" with "tar"? On the other hand, there would likely be a lot of other issues with the rest of the flags too...

Yes, there would probably be more problems if user changed gcc to tar - however the case that may be more problematic is places where people have their own wrapper around GCC. Renesas has something like that.

> Another hackish approach is to check the generated command if it ends with gcc(.exe)? and in that case, enable it.

That is probably ok too.

I am less familiar with this code that you are, so it may be best to solicit other people's input on this. The heavy extenders of this in the open source are Jantje and Liviu, but the vendors like you folk at ST are probably well placed to come to a conclusion on this.

Patch-set: 2
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>