Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> (2 comments)
> 
> Not to make more work for you and I haven't thought this through - but I was wondering if the new public classes can just extend the internal ones and then there would be no duplication or breakage? 

I thought about that, but what will happen when we one day want to get rid of the internal classes as it makes no sense to keep both in the long run? Wouldn't it be a break on the ABI?
We could also do the opposite, but I fear that we yet again will have an ABI incompatibility...

> Anyway, this looks fine to me as is (you can do the String.join in a follow up commit).
> 
> PS this is fine for 10.2.0 too I think.

Ok!

Patch-set: 2