Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > (1 comment)
> > 
> > Do you see any problem doing the "matchConfigurations" call instead of root.getWorkspace().build() call?
> > I'm not sure what will happen with other projects than those of CDT type, but maybe we do not have to care about those? Headless build in CDT is after all focusing on CDT projects...
> 
> Sorry, I don't know either way.

Do you know who could know?

I've been testing this on normal CDT projects and they build the same way as the old implementation, but get the correct verdict.
I can drop a mail to the ML too regarding it and see if someone has an idea...

Patch-set: 1