Update patch set 2
Patch Set 2:

(2 comments)

I do not have any knownledge of the internals of the indexer, so I can't tell if this code is correct or not.

Please also use a shorter initial line in the commit message and then add the details on the lines after. Here are some good tips on how to write good commit messages: https://chris.beams.io/posts/git-commit/

Patch-set: 2
CC: Gerrit User 178588 <178588@97ee7c02-f12f-4043-b43e-dea463d88b31>