Update patch set 1
Patch Set 1:

I had a question from the team if this could be added to ease the use of the headless build for people that are not used to the eclipse command line options (we have a small wrapper script that handles the common eclipse options for selecting the headless build app).

Do you think it's a good thing to add here?

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>