Update patch set 2
Patch Set 2:

If -help works with nothing else (as in how I did it with my comment) then nothing else is needed. I am surprised it works because I thought the resources bundle would start earlier and get in the way as it is a dep of this bundle.

Patch-set: 2
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>