Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > If -help works with nothing else (as in how I did it with my comment) then nothing else is needed. I am surprised it works because I thought the resources bundle would start earlier and get in the way as it is a dep of this bundle.
> > 
> > At least it works in runtime (by clicking "show command line" in the debug configuration UI and then execute that in a terminal, but with -data removed).
> 
> You need -Dosgi.dataAreaRequiresExplicitInit=true which is added (by who??) in eclipse.ini but I think the default if false if not set as a system property to remain compatible.
> 
> You can test from the launch config by putting the -Dosgi.dataAreaRequiresExplicitInit=true in vmargs and leaving the "Location" field blank in the Main tab.

Yes, you are right, doing so fails the launch.
I will create a new bundle like you did for the formatter app, but is there any way to avoid duplicating the help message?
Would it be okay if the HeadlessBuilder class calls a static method in the wrapper class to retrieve the full help message or should I even make the Messages class API-ish?
I think it's error prune to define the text in 2 plugins and would like to avoid that...

Patch-set: 2