Update patch set 4
Patch Set 4:

(1 comment)

> 4 Move the help to the wrapper entirely and have the main class not handle the help at all as it should be unreachable (perhaps in this case the -help in the main class can say something short like refer to the help of the wrapper for the occasional extender like Evgueni Driouk) - the wrapper class can display help on a specific return or exception from the main class
> 5 Move all the argument parsing to the new bundle and then a structure rather than String[] args are passed to main class
> I would do 4 or 5. 4 if I had limited time and 5 if I wanted to end up with a better designed system with a better separation of concerns.

I wanted to do 5, but I don't have enough time to get this into 10.2, so I ended up with 4.
An interesting fact is that you cannot use exceptions defined in the main bundle as they appear to be loaded before the own class is started to be executed.

One thing that I'm uncertain about is if there can be a mix of version of the wrapper bundle and the main bundle so that the help text gets out of sync with the implementation when using 4.
Also, doing 5 all the way would make a clear break of any extenders, although no one should do that for the headless builder class...
Lets do the separation and cleanup of the headless builder a later CDT version.

> I would also include the code that gives nice error messages on missing -data - https://git.eclipse.org/r/c/jdt/eclipse.jdt.core/+/152535/15/org.eclipse.jdt.core.formatterapp/src/org/eclipse/jdt/core/formatterapp/CodeFormatterApplicationWrapper.java#112 - I am not sure anyone really reviewed that complicated catch block, so if you find any issue with it I may pass that up to JDT too.

I debugged it and it looks correct AFAICT.


@Jonah: Can you take a look at this and see if there is anything that I've missed. My limited testing tonight (getting tired) did not show anything unexpected.

Patch-set: 4