Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> @Jonah, I would really like to have this merged for the CDT 10.2 to reduce the pressure from the consumers of our ST product. Can it wait until after RC1 to, maybe, get some comment from the reviewers?

Sorry - I don't have an answer or insight on this area. I don't have time to verify one way or another its suitability.

However if you are confident this is the correct approach, you are a committer so you can put it in RC1 and if we receive feedback in the coming days we can revert for RC2.

HTH.
Jonah

Patch-set: 3
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>