Update patch set 1
Patch Set 1:

I am concerned that we don't encourage users to create relative paths between projects by this mechanism. A project can be located anywhere on the file system and can be located independently of any other project in the workspace. I think it preferable to encourage paths relative to ${workspace_loc} or to add a "linked resource" folder within the project to access external files.

Patch-set: 1
CC: Gerrit User 846 <846@97ee7c02-f12f-4043-b43e-dea463d88b31>