Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I am concerned that we don't encourage users to create relative paths between projects by this mechanism. A project can be located anywhere on the file system and can be located independently of any other project in the workspace. I think it preferable to encourage paths relative to ${workspace_loc} or to add a "linked resource" folder within the project to access external files.
> 
> Thanks for you review John!
> 
> While I mostly agree with you, I find it strange that we allow users to use absolute paths to achieve the same thing.
> I guess you can sort-of achieve the same end result using something like ${workspace_loc:/$ProjName/../foo.elf}, I don't see that as easy as simply writing ../foo.elf. As most users in Eclipse are not into all the different expansion variables, I find it unlikely that the users will find out that this can be done this way without doing extensive documentation of it. It's more natural to simply do ../foo.elf in that case.
> 
> Also, even if we encourage to use other ways, do we have to lock out other ways to achieve the same end result?

Support for absolute paths is necessary to support ${workspace_loc}, ${eclipse_home} and similar variables which must ultimately resolve to an absolute path. But using an absolute path directly within a project is clearly not desirable for reasons of project portability. When we use ${workspace_loc:myProjectName} or ${eclipse_home} we abstract away the underlying absolute location and therefore provide project portability. Using relative paths to files outside a project also destroys portability. The user is forced to import the project at a certain location relative to the file which it references.

Eclipse typically provides a "Variables..." button as an alternative to the "Browse..." button in places where references to files in other projects may be required.

Patch-set: 2