Update patch set 2
Patch Set 2:

As implemented you can't do this. These statements are not equivalent, especially when it comes to links, but also virtual file system cases too:
    IProject.getLocation().append(programPath) 
    IProject().getFile(programPath)
I mean that when programPath is just a path within the project.

I assume that your use case is programPath = "../bin/a.elf" or similar, in that case it won't make a difference because of the leading ".."

Not sure how to handle the general case on this. Could we try the old version first (getFile) and if a file is not found there, then try getLocation().append(...). The box in the launch configuration is already checked against absolute files and project files.

--

I was going to recommend the special case can just be the programPath starts with ".." -- but getFile() does actually handle ".." between two projects today. e.g. if you have two projects like this:

/tmp/proj/p1 and /tmp/proj/p2 and p1 has proj.elf in it.

Then this works:

   IProject project = ResourcesPlugin.getWorkspace().getRoot().getProject("p2");
   IFile file = project.getFile("../p1/proj.elf");
   file.getProject().getName() will return p1

--

I think it is too late for 10.2 for a change like this.

Patch-set: 2
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>